# backend/docs/openapi_spec.yaml
openapi: 3.0.3
info:
  title: Revive AI API
  description: |
    AI-Powered Reputation Intelligence & Customer Recovery Platform
    
    ## Overview
    Revive AI helps businesses monitor public reviews, analyze customer conversations,
    predict churn risks, and take automated recovery actions to improve ratings,
    retention, and revenue.
    
    ## Authentication
    All endpoints require JWT Bearer token authentication (except public endpoints).
    Include the token in the Authorization header: `Bearer <token>`
  version: 1.0.0
  contact:
    name: Revive AI Support
    email: support@revive-ai.com
    url: https://revive-ai.com/support
  license:
    name: Proprietary License
    url: https://revive-ai.com/license
  termsOfService: https://revive-ai.com/terms
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://staging-api.revive-ai.com
    description: Staging server
  - url: https://api.revive-ai.com
    description: Production server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
  schemas:
...
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        platform:
          type: string
          enum: [google, yelp, facebook, trustpilot]
        external_id:
          type: string
        customer_name:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        content:
          type: string
        sentiment_score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        urgency_level:
          type: string
          enum: [low, medium, high]
        issue_categories:
          type: array
          items:
            type: string
            enum: [support, pricing, delivery, quality]
        status:
          type: string
          enum: [pending, responded, escalated]
        requires_private_recovery:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ReviewIngest:
      type: object
      required:
        - platform
        - rating
        - content
      properties:
        platform:
          type: string
        external_id:
          type: string
        customer_name:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        content:
          type: string
        created_at:
          type: string
          format: date-time
...
paths:
...
  /api/v1/reviews/ingest:
    post:
      tags:
        - reviews
      summary: Ingest new review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewIngest'
      responses:
        '201':
          description: Review ingested successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  review_id:
                    type: string
                  sentiment_score:
                    type: number
                  urgency_level:
                    type: string
                  issue_categories:
                    type: array
                    items:
                      type: string
  /api/v1/reviews:
    get:
      tags:
        - reviews
      summary: List reviews
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: rating
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: urgency_level
          in: query
          schema:
            type: string
            enum: [low, medium, high]
      responses:
        '200':
          description: List of reviews
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
...